requirements.txt
pymongo[srv]==3.12
flask-pymongo
bcrypt
dotenv